Add unit tests for Basic, ShoppingCart, and BranchManagement functionalities
