+++ tests/Pure.Domain.Test/BasicTests.cs
@@ -1,0 +1,31 @@
+using Pure.Domain.Generated;
+using Sekiban.Pure.Projectors;
+using Sekiban.Pure.Documents;
+namespace Pure.Domain.Test;
+
+public class BasicTests
+{
+    [Fact]
+    public void Test1()
+    {
+        var version = GetVersion<UserProjector>();
+        Assert.Equal("1.0.1", version);
+    }
+    [Fact]
+    public void PartitionKeysTest()
+    {
+        var partitionKeys = PartitionKeys.Generate();
+        Assert.Equal(PartitionKeys.DefaultAggregateGroupName, partitionKeys.Group);
+        Assert.Equal(PartitionKeys.DefaultRootPartitionKey, partitionKeys.RootPartitionKey);
+    }
+    [Fact]
+    public void TenantPartitionKeysTest()
+    {
+        var partitionKeys = TenantPartitionKeys.Tenant("tenant1").Generate("group1");
+        Assert.Equal("tenant1", partitionKeys.RootPartitionKey);
+        Assert.Equal("group1", partitionKeys.Group);
+    }
+    
+    public string GetVersion<TAggregateProjector>() where TAggregateProjector : IAggregateProjector, new() =>
+        new TAggregateProjector().GetVersion();
+}
